requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==0.12.0
    # via
    #   tensorboard
    #   tensorflow
aiohttp==3.7.4.post0
    # via fsspec
astunparse==1.6.3
    # via tensorflow
async-timeout==3.0.1
    # via aiohttp
attrs==20.3.0
    # via aiohttp
blis==0.7.4
    # via
    #   spacy
    #   thinc
cachetools==4.2.1
    # via google-auth
catalogue==2.0.1
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via
    #   aiohttp
    #   requests
click==7.1.2
    # via typer
cymem==2.0.5
    # via
    #   preshed
    #   spacy
    #   thinc
flatbuffers==1.12
    # via tensorflow
fsspec[http]==0.8.7
    # via pytorch-lightning
future==0.18.2
    # via pytorch-lightning
gast==0.3.3
    # via tensorflow
google-auth-oauthlib==0.4.3
    # via tensorboard
google-auth==1.27.1
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.32.0
    # via
    #   tensorboard
    #   tensorflow
h5py==2.10.0
    # via tensorflow
idna==2.10
    # via
    #   requests
    #   yarl
jinja2==2.11.3
    # via spacy
joblib==1.0.1
    # via scikit-learn
keras-preprocessing==1.1.2
    # via tensorflow
markdown==3.3.4
    # via tensorboard
markupsafe==1.1.1
    # via jinja2
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
murmurhash==1.0.5
    # via
    #   preshed
    #   spacy
    #   thinc
numpy==1.19.5
    # via
    #   blis
    #   h5py
    #   keras-preprocessing
    #   opt-einsum
    #   pytorch-lightning
    #   scikit-learn
    #   scipy
    #   seqeval
    #   spacy
    #   tensorboard
    #   tensorflow
    #   thinc
    #   torch
oauthlib==3.1.0
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
packaging==20.9
    # via spacy
pathy==0.4.0
    # via spacy
preshed==3.0.5
    # via
    #   spacy
    #   thinc
protobuf==3.15.6
    # via
    #   tensorboard
    #   tensorflow
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pydantic==1.7.3
    # via
    #   spacy
    #   thinc
pyparsing==2.4.7
    # via packaging
pytorch-lightning==1.2.3
    # via -r requirements.in
pyyaml==5.3.1
    # via pytorch-lightning
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
requests==2.25.1
    # via
    #   fsspec
    #   requests-oauthlib
    #   smart-open
    #   spacy
    #   tensorboard
rsa==4.7.2
    # via google-auth
scikit-learn==0.24.1
    # via seqeval
scipy==1.6.1
    # via scikit-learn
seqeval==1.2.2
    # via -r requirements.in
six==1.15.0
    # via
    #   absl-py
    #   astunparse
    #   google-auth
    #   google-pasta
    #   grpcio
    #   h5py
    #   keras-preprocessing
    #   protobuf
    #   tensorboard
    #   tensorflow
smart-open==3.0.0
    # via pathy
spacy-legacy==3.0.1
    # via spacy
spacy==3.0.5
    # via -r requirements.in
srsly==2.4.0
    # via
    #   spacy
    #   thinc
tensorboard-plugin-wit==1.8.0
    # via tensorboard
tensorboard==2.4.1
    # via
    #   pytorch-lightning
    #   tensorflow
tensorflow-estimator==2.4.0
    # via tensorflow
tensorflow==2.4.1
    # via -r requirements.in
termcolor==1.1.0
    # via tensorflow
thinc==8.0.2
    # via spacy
threadpoolctl==2.1.0
    # via scikit-learn
torch==1.7.1
    # via
    #   -r requirements.in
    #   pytorch-lightning
tqdm==4.56.2
    # via
    #   -r requirements.in
    #   pytorch-lightning
    #   spacy
typer==0.3.2
    # via
    #   pathy
    #   spacy
typing-extensions==3.7.4.3
    # via
    #   aiohttp
    #   tensorflow
    #   torch
urllib3==1.26.3
    # via requests
wasabi==0.8.2
    # via
    #   spacy
    #   thinc
werkzeug==1.0.1
    # via tensorboard
wheel==0.36.2
    # via
    #   astunparse
    #   tensorboard
    #   tensorflow
wrapt==1.12.1
    # via tensorflow
yarl==1.6.3
    # via aiohttp
wandb==0.12.10
# The following packages are considered to be unsafe in a requirements file:
# setuptools
